Kill the process to end a lingering wakelock
Problem is that when you disable the service, it doesn't properly die. Whatever wakelock was started by SystemUpdateService persists as long as com.google.android.gms process is running. So we kill the process precisely once.

The RootShell library acts funny when you are entering multiple commands. For whatever reason, they are not all executed, and not in actual order, and there is no way to sleep it between them because its all sharing a single process. So we do the first step of disabling the component com.google.android.gms/.updater.SystemUpdateService, and then create an alarm for 20 seconds to launch the OnAlarmReceiver, which checks in the output of "dumpsys power" for the SystemUpdateService wakelock, and if it finds it, kills the com.google.android.gms process.