Kill the receivers
Seems that the SystemUpdateService contains a number of private classes, which are receivers for certain system events, like network connectivity changes. When it receives that notice, it creates a wake lock, which prevents the device from sleeping, and since SystemUpdateService was disabled, the wakelock would never be cleared. This is very bad. Means that even killing the gps processess after disabling the SystemUpdateService, the wakelock could still be recreated when network connectivity changes.

Now it cannot, because we also disable the receiver that is responsible for creating the wakelock in the first place.