Request a root authorization from BullshitActivity
The process of wasting one reboot just to try to catch an unreliable root authorization request was annoying, so we surround the self termination line inside BullshitActivity with a root authorization request. Now when the activity is started for the purpose of satisfying the requirement for manually running something in the package in order to enable the receivers, we also satisfy the root authorization request.